Refactor: split SheetViewer, FretboardScreen, StrumPatternsView
Pure file extraction — no logic changes.

SheetViewer.kt (979→855): extract LabelDisplayRow and SongStatsRow
to SheetViewerComponents.kt

FretboardScreen.kt (882→759): extract NAV_* constants, data classes,
drawerSections(), and navigateToChord() to NavigationConstants.kt

StrumPatternsView.swift (775→561): extract CreateStrumPatternSheet,
CreateFingerpickPatternSheet, and utility functions to separate files.
Helper functions renamed with pattern* prefix to avoid collisions.
